+++ core/processor/abci.go
@@ -134,6 +134,10 @@ type ReferralProgram interface {
 	ApplyReferralCode(context.Context, types.PartyID, types.ReferralSetID) error
 }
 
+type VolumeDiscountProgram interface {
+	UpdateProgram(program *types.VolumeDiscountProgram)
+}
+
 type BlockchainClient interface {
 	Validators(height *int64) ([]*tmtypesint.Validator, error)
 }
@@ -201,6 +205,7 @@ type App struct {
 	stateVar               StateVarEngine
 	teamsEngine            TeamsEngine
 	referralProgram        ReferralProgram
+	volumeDiscountProgram  VolumeDiscountProgram
 	protocolUpgradeService ProtocolUpgradeService
 	erc20MultiSigTopology  ERC20MultiSigTopology
 	gastimator             *Gastimator
@@ -244,6 +249,7 @@ func NewApp(
 	stateVarEngine StateVarEngine,
 	teamsEngine TeamsEngine,
 	referralProgram ReferralProgram,
+	volumeDiscountProgram VolumeDiscountProgram,
 	blockchainClient BlockchainClient,
 	erc20MultiSigTopology ERC20MultiSigTopology,
 	version string, // we need the version for snapshot reload
@@ -293,6 +299,7 @@ func NewApp(
 		stateVar:               stateVarEngine,
 		teamsEngine:            teamsEngine,
 		referralProgram:        referralProgram,
+		volumeDiscountProgram:  volumeDiscountProgram,
 		version:                version,
 		blockchainClient:       blockchainClient,
 		erc20MultiSigTopology:  erc20MultiSigTopology,
@@ -1920,6 +1927,8 @@ func (app *App) onTick(ctx context.Context, t time.Time) {
 			app.enactMarketStateUpdate(ctx, prop)
 		case toEnact.IsReferralProgramUpdate():
 			app.referralProgram.UpdateProgram(toEnact.ReferralProgramUpdate())
+		case toEnact.IsVolumeDiscountProgramUpdate():
+			app.volumeDiscountProgram.UpdateProgram(toEnact.VolumeDiscountProgramUpdate())
 		default:
 			app.log.Error("unknown proposal cannot be enacted", logging.ProposalID(prop.ID))
 			prop.FailUnexpectedly(fmt.Errorf("unknown proposal \"%s\" cannot be enacted", prop.ID))